+++ RELEASE_NOTES.md
@@ -5,6 +5,7 @@
 * GMB-1116 - Replicate remote uploader files when it is using file system
 * GMB-1119 - Lifetime Brands LENS Status Research
 * GMB-1121 - Trigger file is not replicated on successful staging.
+* GSB-3244: setEngineStateToError exposes Message Content in LENS Logs.
 
 5.2.13
 ---------------